Data.TypeMergingList

Description

A Snoc List type that merges elements of the same constructor using the Semigroup instance. For example, suppose we have some data type: > data Entry = IntEntry Int | StringEntry String deriving (Eq, Data, Show) > instance Semigroup Entry where > IntEntry a <> IntEntry b = IntEntry (a + b) > StringEntry a <> StringEntry b = StringEntry (a ++ b) > _ <> b = b -- if the constructors don't match, just take the right one

Then we can do: > snoc (TypeMergingList []) (IntEntry 1) = TypeMergingList [IntEntry 1] > snoc (TypeMergingList [IntEntry 1]) (IntEntry 2) = TypeMergingList [IntEntry 3] > snoc (TypeMergingList [IntEntry 1]) (StringEntry "hello") = TypeMergingList [IntEntry 1, StringEntry "hello"] > snoc (TypeMergingList [IntEntry 1, StringEntry "hello"]) (StringEntry "world") = TypeMergingList [IntEntry 1, StringEntry "helloworld"]

Documentation

newtype TypeMergingList a

Constructors

TypeMergingList [a]

class Data a => DataMergeable a where

Class of partially mergeable types. Instances of this class may assume that the constructors of the two arguments are the same (i.e. toConstr x == toConstr y), and are permitted to be partial if this is not the case.

Methods

merge :: a -> a -> a

errorDifferentConstructors :: Data a => a -> a -> b

getByCtor :: forall (ctor :: Symbol) s a. (Generic s, AsConstructor ctor s s a a) => TypeMergingList s -> Maybe a

snoc :: DataMergeable a => TypeMergingList a -> a -> TypeMergingList a

append :: DataMergeable a => TypeMergingList a -> TypeMergingList a -> TypeMergingList a

fromList :: (DataMergeable a, Data a) => [a] -> TypeMergingList a

toList :: TypeMergingList a -> [a]

toVector :: TypeMergingList a -> Vector a