JVM.Data.Abstract.Builder.Label

Description

Label type for instructions This data type allows referring to specific instructions (for jumps) without having to manually calculate their offsets. A label represents a unique instruction in a Code attribute, and is marked by inserting a Label instruction directly above the instruction it refers to. For example, the following (contrived) code:

       0: aload_0
       1: ifeq          4
       4: return

could be written as follows, using CodeBuilder:

emit ALoad0
label <- newLabel
emit (IfEq label)
emit (Label label)
emit Return

Documentation

newtype Label

Constructors

MkLabel Int