JVM.Data.Abstract.ConstantPool

Description

Provides a monadic interface to the constant pool. This aims to eliminate the need to manually specify the index of the constant

Documentation

data ConstantPoolEntry

High-level, type-safe representation of a constant pool entry This tries to hide indexes as much as possible, instead just allowing the values to be provided directly. These are transformed into the correct indexes when the constant pool is built, which uses a state monad to avoid repeating entries.

Constructors

CPClassEntry
A class reference

Fields

ClassInfoType

The class being referenced

CPFieldRefEntry FieldRef

CPMethodRefEntry MethodRef

CPInterfaceMethodRefEntry MethodRef

CPStringEntry Text

CPIntegerEntry Int

CPFloatEntry Float

CPLongEntry Int64

CPDoubleEntry Double

CPNameAndTypeEntry Text Text

CPUTF8Entry Text

CPMethodHandleEntry MethodHandleEntry

CPMethodTypeEntry MethodDescriptor

CPInvokeDynamicEntry
CONSTANT_InvokeDynamic_info

Fields

BootstrapMethod

bootstrap_method_attr(_index)

Text

name(_and_type_index)

MethodDescriptor

(name_and_)type(_index)

data MethodHandleEntry

Constructors

MHGetField FieldRef

MHGetStatic FieldRef

MHPutField FieldRef

MHPutStatic FieldRef

MHInvokeVirtual MethodRef

MHNewInvokeSpecial MethodRef

MHInvokeStatic MethodRef

MHInvokeSpecial MethodRef

MHInvokeInterface MethodRef

data FieldRef

Constructors

FieldRef ClassInfoType Text FieldType

data MethodRef

Constructors

MethodRef

Fields

ClassInfoType

The class containing the method

Text

The name of the method

MethodDescriptor

The descriptor of the method

data BootstrapMethod

Constructors

BootstrapMethod MethodHandleEntry [BootstrapArgument]

data BootstrapArgument

Constructors

BMClassArg ClassInfoType

BMStringArg Text

BMIntArg Int

BMMethodArg MethodDescriptor

BMMethodHandleArg MethodHandleEntry

bmArgToCPEntry :: BootstrapArgument -> ConstantPoolEntry