JVM.Data.Abstract.Instruction

Description

High level representation of a JVM instruction, with type-safe arguments and no stack manipulation needed. This is not a 1-1 mapping to the actual instructions, use Instruction for that. Instead, this includes some conveniences like the Label instruction to avoid manually dealing with jump offsets.

Documentation

type Reference = Int

type Instruction = Instruction' Label

data Instruction' label

Constructors

ALoad U1

AStore U1

AReturn

AConstNull

Dup

IfEq label

IfNe label

IfLt label

IfGe label

IfGt label

IfLe label

Instanceof ClassInfoType

InvokeStatic ClassInfoType Text MethodDescriptor

InvokeInterface ClassInfoType Text MethodDescriptor

InvokeVirtual ClassInfoType Text MethodDescriptor

InvokeDynamic BootstrapMethod Text MethodDescriptor

InvokeSpecial ClassInfoType Text MethodDescriptor

ILoad U1

IStore U1

Label label

LDC LDCEntry

PutStatic ClassInfoType Text FieldType

GetField ClassInfoType Text FieldType

GetStatic ClassInfoType Text FieldType

PutField ClassInfoType Text FieldType

Goto label

CheckCast ClassInfoType

Return

New ClassInfoType

jumpTarget :: Instruction' label -> Maybe label

data LDCEntry

Constructors

LDCInt Int

LDCFloat Float

LDCString Text

LDCClass ClassInfoType

ldcEntryToFieldType :: LDCEntry -> FieldType