JVM.Data.Abstract.Name

Documentation

newtype PackageName

A JVM package name This is defined as a potentially empty list of identifiers, which would be separated by dots in the source code

Constructors

PackageName [Text]

parsePackageName :: Text -> PackageName

Parse a PackageName from a Text

>>> parsePackageName "java.lang"
PackageName ["java","lang"]

>>> parsePackageName ""
PackageName []

newtype ClassName

A JVM class name

Constructors

ClassName Text

parseClassName :: Text -> ClassName

Parse a ClassName from a Text

data QualifiedClassName

Constructors

QualifiedClassName PackageName ClassName

parseQualifiedClassName :: Text -> QualifiedClassName

Parse a QualifiedClassName from a Text

>>> parseQualifiedClassName "java.lang.Object"
QualifiedClassName (PackageName ["java","lang"]) (ClassName "Object")

>>> parseQualifiedClassName "Object"
QualifiedClassName (PackageName []) (ClassName "Object")

This function is lenient and will accept invalid class names: >>> parseQualifiedClassName "123invalid" QualifiedClassName (PackageName []) (ClassName "123invalid")

toInternalName :: QualifiedClassName -> Text

Convert a QualifiedClassName to a Text that can be used as an internal JVM name

>>> toInternalName (QualifiedClassName (PackageName ["java","lang"]) (ClassName "Object"))
"java/lang/Object"

suitableFilePath :: QualifiedClassName -> FilePath

Convert a QualifiedClassName to a FilePath that one would expect the class file to be in >>> suitableFilePath "java.lang.Object" "javalangObject.class"

>>> suitableFilePath "Object"
"Object.class"