JVM.Data.Analyse.StackMap

Description

Generate a stack map table for a method. This process MUST run last in the high level stage - modifications to the code after this point will invalidate the stack map table and cause invalid class files to be generated.

Documentation

data BasicBlock

Constructors

BasicBlock

Fields

index :: Int

instructions :: [Instruction]

end :: Maybe Label

data Frame

Constructors

Frame

Fields

locals :: [LocalVariable]

stack :: [StackEntry]

data LocalVariable

Constructors

Uninitialised

LocalVariable FieldType

data StackEntry

Constructors

StackEntry FieldType

StackEntryTop

StackEntryNull

lvToStackEntry :: LocalVariable -> StackEntry

stackEntryToLV :: StackEntry -> LocalVariable

splitIntoBasicBlocks :: [Instruction] -> [BasicBlock]

splitOnLabels :: [Instruction] -> [(Maybe Label, [Instruction])]

topFrame :: MethodDescriptor -> Frame

analyseBlockDiff :: Frame -> BasicBlock -> Frame

frameDiffToSMF :: HasCallStack => Frame -> BasicBlock -> StackMapFrame

lvToVerificationTypeInfo :: LocalVariable -> VerificationTypeInfo

seToVerificationTypeInfo :: StackEntry -> VerificationTypeInfo

calculateStackMapFrames :: MethodDescriptor -> [Instruction] -> [StackMapFrame]

replaceAtOrGrow :: Int -> LocalVariable -> [LocalVariable] -> [LocalVariable]

replaceAt :: Int -> a -> [a] -> [a]

takeWhileInclusive :: (a -> Bool) -> [a] -> [a]