JVM.Data.Raw.AccessFlags

Documentation

data AccessFlag

Very thin wrapper around the access flags defined in the JVM spec. These flags may be for a class, field, or method - for a safer interface, see AccessFlags.

Constructors

ACC_PUBLIC

ACC_PRIVATE

ACC_PROTECTED

ACC_STATIC

ACC_FINAL

ACC_VOLATILE

ACC_TRANSIENT

ACC_SUPER

ACC_INTERFACE

ACC_ABSTRACT

ACC_SYNTHETIC

ACC_ANNOTATION

ACC_ENUM

ACC_SYNCHRONIZED

ACC_BRIDGE

ACC_VARARGS

ACC_NATIVE

ACC_STRICT

accessFlagValue :: AccessFlag -> Word16