JVM.Data.Raw.ClassFile

Documentation

data ClassFile

Low level representation of the class file format. This is the closest representation to the actual format of the class file, that makes only a few compromises for convenience, such as omitting the explicit x_count fields, instead calculating them from the length of the corresponding vectors. This has the added bonus of making it impossible to construct an invalid class file (in this regard anyway)

Constructors

ClassFile

Fields

magic :: Word32

minorVersion :: MinorVersion

majorVersion :: MajorVersion

constantPool :: Vector ConstantPoolInfo

accessFlags :: Word16

thisClass :: Word16

superClass :: Word16

interfaces :: Vector Word16

fields :: Vector FieldInfo

methods :: Vector MethodInfo

attributes :: Vector AttributeInfo

data FieldInfo

Constructors

FieldInfo

Fields

accessFlags :: Word16

nameIndex :: Word16

descriptorIndex :: Word16

attributes :: Vector AttributeInfo

data AttributeInfo

Constructors

AttributeInfo

Fields

nameIndex :: Word16

info :: Attribute

data Attribute

Constructors

CodeAttribute

Fields

maxStack :: Word16

maxLocals :: Word16

code :: Vector Instruction

exceptionTable :: Vector ExceptionTableEntry

codeAttributes :: Vector AttributeInfo

InnerClassesAttribute (Vector InnerClassInfo)

LineNumberTableAttribute (Vector LineNumberTableEntry)

StackMapTableAttribute (Vector StackMapFrame)

ConstantValueAttribute Word16

SourceFileAttribute Word16

BootstrapMethodsAttribute (Vector BootstrapMethod)

data InnerClassInfo

Constructors

InnerClassInfo

Fields

innerClassInfo :: Word16

outerClassInfo :: Word16

innerName :: Word16

accessFlags :: Word16

data LineNumberTableEntry

Constructors

LineNumberTableEntry

Fields

startPc :: U2

lineNumber :: U2

data StackMapFrame

The stack map frame types are defined as follows: > union stack_map_frame { > same_frame; > same_locals_1_stack_item_frame; > same_locals_1_stack_item_frame_extended; > chop_frame; > same_frame_extended; > append_frame; > full_frame; >}

Constructors

SameFrame U1
0-63

SameLocals1StackItemFrame VerificationTypeInfo U1
      same_locals_1_stack_item_frame {
         u1 frame_type = SAME_LOCALS_1_STACK_ITEM; /* 64-127 */
         verification_type_info stack[1];
     }

SameLocals1StackItemFrameExtended VerificationTypeInfo U2

SameFrameExtended U2

ChopFrame U1 U2

AppendFrame (Vector VerificationTypeInfo) U2

FullFrame (Vector VerificationTypeInfo) (Vector VerificationTypeInfo) U2

data VerificationTypeInfo

Constructors

TopVariableInfo

IntegerVariableInfo

FloatVariableInfo

LongVariableInfo

DoubleVariableInfo

NullVariableInfo

UninitializedThisVariableInfo

ObjectVariableInfo Word16

UninitializedVariableInfo U2

data BootstrapMethod

Constructors

BootstrapMethod

Fields

bootstrapMethodRef :: Word16

bootstrapArguments :: Vector Word16

data ExceptionTableEntry

Constructors

ExceptionTableEntry

Fields

startPc :: Word16

endPc :: Word16

handlerPc :: Word16

catchType :: Word16

data MethodInfo

Constructors

MethodInfo

Fields

accessFlags :: Word16

nameIndex :: Word16

descriptorIndex :: Word16

attributes :: Vector AttributeInfo

putConstantPool :: Vector ConstantPoolInfo -> Put

putInterfaceTable :: Vector Word16 -> Put

putFieldTable :: Vector FieldInfo -> Put

putAttributes :: Vector AttributeInfo -> Put

putAttribute :: Attribute -> Put

Puts the attribute data. This must not include length or name, just the actual value

putBootstrapMethod :: BootstrapMethod -> Put

putExceptionTable :: Vector ExceptionTableEntry -> Put

putMethodTable :: Vector MethodInfo -> Put